# koan/openapi.yaml
...
components:
  schemas:
    Error:
      properties:
        error:
          properties:
            code:
              type: string
            message:
              type: string
          required:
          - code
          - message
          type: object
      required:
      - error
      type: object
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
info:
  description: Optional, token-authenticated HTTP control layer for the Kōan agent. See docs/operations/rest-api.md.
  title: Kōan REST API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/logs:
    get:
      operationId: observability_logs_get
      parameters:
      - description: Log source to read.
        in: query
        name: source
        required: false
        schema:
          default: all
          enum:
          - run
          - awake
          - all
          type: string
      - description: Maximum lines returned per source.
        in: query
        name: limit
        required: false
        schema:
          default: 200
          maximum: 2000
          minimum: 1
          type: integer
      - description: Case-insensitive substring filter.
        in: query
        name: q
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Tail recent agent logs, optionally filtered by source.
      tags:
      - observability
...
  /v1/usage:
    get:
      operationId: observability_usage_get
      parameters:
      - description: Window length; values are clamped to the documented range.
        in: query
        name: days
        required: false
        schema:
          default: 7
          maximum: 100
          minimum: 1
          type: integer
      - description: Shift the window back by this many granularity units.
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      - description: Series bucketing granularity.
        in: query
        name: granularity
        required: false
        schema:
          default: day
          enum:
          - day
          - week
          - month
          type: string
      - description: Include a per-project series breakdown.
        in: query
        name: stacked
        required: false
        schema:
          default: false
          type: boolean
      - description: Restrict totals and series to one project.
        in: query
        name: project
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Daily agent usage totals, optionally split by project.
      tags:
      - observability
security:
- bearerAuth: []
servers:
- description: Default loopback bind
  url: http://127.0.0.1:8420
...